refactor: moved remaining classes to transactionhandlers package
